backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0       # required by pymysql for SSL / caching_sha2_password
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9    # form data / file uploads
openai>=1.14.0
httpx>=0.27.0
python-dotenv>=1.0.0
